Simplify the component usage
